Adds a few initial unit tests
The refactoring done so far has been minimally invasive, and testing
that the executable builds and runs without any obvious problems has
been the only verification done to get some confidence that we didn't
introduce any bugs.

This, of course, is not a very reliable way to do refactoring, so this
change introduces the first few tests for the project in preparation for
more 'aggresive' refactoring.

We're using the Catch project (https://github.com/philsquared/Catch/)
due to its simplicity (it requires the inclusion of a single header file
only) and features. One drawback I saw was that compilation is somewhat
slow for the main test file, but not slow enough to consider it a deal-breaker.