[TEST] Improves count_set_bits test
These tests (along with some others related to bit twiddling) are a bit
opaque because of the magic numbers that hide the binary pattern we're
testing, so adding appropriate comments is a must.

When we migrate to C++14, we'll be able to get rid of those comments and
use binary literals instead.