Refactor statistics out of AlphaBetaSearch
While refactoring, I noticed that the statistics seemed way off, so I
discovered there was a major bug in the implementation of the iterative
deepening search. I don't remember that bug having been present years
ago when it was first written, so it is possible it slipped in during a
later refactoring (the perils of not having tests!)

I took the opportunity to update the compilation to use C++14, and
fortunately nothing broke.