:hammer: Extract `bits` from `util` ns
The `util` namespace had become a wildcard namespace where we put things
indiscriminately. Obviously, this was a terrible idea. The most obvious
reason why is that a lot of classes had to be recompiled often for no
good reason (e.g., a change in a bitboard routine would end up forcing a
recompilation of the GeneticAlgorithm class)

This commit extracts a new namespace `bits::`, removing a lot of
unnecessary dependencies in a lot of files (hence the large number of
modified files besides the ones directly affected by the change in
namespace of routines like `msb_position`)

While I was a it, I took the opportunity to see how fast some of the bit
manipulation routines were. We can now query that information by
running:

> make perf_test

I also extended some of the unit tests we had to add a bit of additional
coverage for some of those bit-twiddling routines.